refactor: fold MCP App support into union/intersect/difference tools
Polygon ops now render as a live Mapbox GL JS map directly from the
existing offline tools instead of via parallel `*_app_tool` siblings.

- Extract the rendering pipeline into a shared `renderPolygonOpsAppHtml`
  module so the MCP Apps resource and the inline MCP-UI rawHtml block use
  one template and one source of truth.
- Slim down `PolygonOpsAppUIResource` to a thin wrapper around the shared
  render function (only the empty/postMessage variant).
- Add `_meta.ui.resourceUri` to `union_tool`, `intersect_tool`, and
  `difference_tool` pointing to the shared `ui://mapbox/polygon-ops-app/
  index.html` resource (MCP Apps path).
- Gate inline MCP-UI rawHtml on `isMcpUiEnabled()` and bake both inputs
  and the result polygon into the iframe with operation-keyed result
  coloring (green=union, purple=intersect, orange=difference).
- Polygon ops are fully offline, so the inline path uses
  `MAPBOX_PUBLIC_TOKEN` from env directly rather than the tokens API.
