+++ src/tools/intersect-tool/IntersectTool.ts
@@ -1,8 +1,10 @@
 // Copyright (c) Mapbox, Inc.
 // Licensed under the MIT License.
 
+import { randomUUID } from 'node:crypto';
 import { intersect, polygon, featureCollection } from '@turf/turf';
 import { context, SpanStatusCode, trace } from '@opentelemetry/api';
+import { createUIResource } from '@mcp-ui/server';
 import { createLocalToolExecutionContext } from '../../utils/tracing.js';
 import { BaseTool } from '../BaseTool.js';
 import { IntersectInputSchema } from './IntersectTool.input.schema.js';
@@ -11,6 +13,8 @@ import {
   type IntersectOutput
 } from './IntersectTool.output.schema.js';
 import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
+import { isMcpUiEnabled } from '../../config/toolConfig.js';
+import { tryRenderPolygonOpsInlineHtml } from '../../resources/ui-apps/polygonOpsAppHtml.js';
 
 export class IntersectTool extends BaseTool<
   typeof IntersectInputSchema,
@@ -29,6 +33,16 @@ export class IntersectTool extends BaseTool<
     destructiveHint: false,
     idempotentHint: true,
     openWorldHint: false
+  };
+
+  readonly meta = {
+    ui: {
+      resourceUri: 'ui://mapbox/polygon-ops-app/index.html',
+      csp: {
+        connectDomains: ['https://*.mapbox.com', 'https://events.mapbox.com'],
+        resourceDomains: ['https://api.mapbox.com']
+      }
+    }
   };
 
   constructor() {
@@ -63,11 +77,44 @@ export class IntersectTool extends BaseTool<
             ? `The polygons intersect.\nIntersection geometry:\n${JSON.stringify(validated.geometry, null, 2)}`
             : 'The polygons do not intersect.';
 
+          const content: CallToolResult['content'] = [
+            { type: 'text' as const, text }
+          ];
+
+          if (isMcpUiEnabled()) {
+            const inlineHtml = tryRenderPolygonOpsInlineHtml({
+              operation: 'intersect',
+              inputs: [poly1, poly2] as Array<{
+                type: 'Feature';
+                geometry: unknown;
+              }>,
+              result: (result ?? null) as {
+                type: 'Feature';
+                geometry: unknown;
+              } | null,
+              summary: validated.intersects
+                ? 'Intersection of two polygons'
+                : 'Polygons do not intersect'
+            });
+            if (inlineHtml) {
+              content.push(
+                createUIResource({
+                  uri: `ui://mapbox/polygon-ops/${randomUUID()}`,
+                  content: { type: 'rawHtml', htmlString: inlineHtml },
+                  encoding: 'text',
+                  uiMetadata: {
+                    'preferred-frame-size': ['100%', '500px']
+                  }
+                })
+              );
+            }
+          }
+
           toolContext.span.setStatus({ code: SpanStatusCode.OK });
           toolContext.span.end();
 
           return {
-            content: [{ type: 'text' as const, text }],
+            content,
             structuredContent: validated,
             isError: false
           };
