+++ src/tools/union-tool/UnionTool.ts
@@ -1,8 +1,10 @@
 // Copyright (c) Mapbox, Inc.
 // Licensed under the MIT License.
 
+import { randomUUID } from 'node:crypto';
 import { union, polygon, featureCollection } from '@turf/turf';
 import { context, SpanStatusCode, trace } from '@opentelemetry/api';
+import { createUIResource } from '@mcp-ui/server';
 import { createLocalToolExecutionContext } from '../../utils/tracing.js';
 import { BaseTool } from '../BaseTool.js';
 import { UnionInputSchema } from './UnionTool.input.schema.js';
@@ -11,6 +13,8 @@ import {
   type UnionOutput
 } from './UnionTool.output.schema.js';
 import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
+import { isMcpUiEnabled } from '../../config/toolConfig.js';
+import { tryRenderPolygonOpsInlineHtml } from '../../resources/ui-apps/polygonOpsAppHtml.js';
 
 export class UnionTool extends BaseTool<
   typeof UnionInputSchema,
@@ -29,6 +33,16 @@ export class UnionTool extends BaseTool<
     destructiveHint: false,
     idempotentHint: true,
     openWorldHint: false
+  };
+
+  readonly meta = {
+    ui: {
+      resourceUri: 'ui://mapbox/polygon-ops-app/index.html',
+      csp: {
+        connectDomains: ['https://*.mapbox.com', 'https://events.mapbox.com'],
+        resourceDomains: ['https://api.mapbox.com']
+      }
+    }
   };
 
   constructor() {
@@ -60,11 +74,36 @@ export class UnionTool extends BaseTool<
             `Result type: ${validated.type}\n` +
             `GeoJSON geometry:\n${JSON.stringify(validated.geometry, null, 2)}`;
 
+          const content: CallToolResult['content'] = [
+            { type: 'text' as const, text }
+          ];
+
+          if (isMcpUiEnabled()) {
+            const inlineHtml = tryRenderPolygonOpsInlineHtml({
+              operation: 'union',
+              inputs: polys as Array<{ type: 'Feature'; geometry: unknown }>,
+              result: merged as { type: 'Feature'; geometry: unknown },
+              summary: `Union of ${input.polygons.length} polygons`
+            });
+            if (inlineHtml) {
+              content.push(
+                createUIResource({
+                  uri: `ui://mapbox/polygon-ops/${randomUUID()}`,
+                  content: { type: 'rawHtml', htmlString: inlineHtml },
+                  encoding: 'text',
+                  uiMetadata: {
+                    'preferred-frame-size': ['100%', '500px']
+                  }
+                })
+              );
+            }
+          }
+
           toolContext.span.setStatus({ code: SpanStatusCode.OK });
           toolContext.span.end();
 
           return {
-            content: [{ type: 'text' as const, text }],
+            content,
             structuredContent: validated,
             isError: false
           };
